requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
openai==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
chromadb==0.4.18
pydantic==2.4.2